+++ spec/requests/admin_timeout_spec.rb
@@ -1,7 +1,7 @@
 require "rails_helper"
 
 RSpec.describe "Admin session timing out", type: :request do
-  let(:timeout_length_in_minutes) { ApplicationController::ADMIN_TIMEOUT_LENGTH_IN_MINUTES }
+  let(:timeout_length_in_minutes) { Admin::BaseAdminController::ADMIN_TIMEOUT_LENGTH_IN_MINUTES }
   let(:user_id) { "userid-345" }
   let(:organisation_id) { "organisationid-6789" }
 
@@ -27,26 +27,6 @@ RSpec.describe "Admin session timing out", type: :request do
 
         follow_redirect!
         expect(response.body).to include("Your session has timed out due to inactivity")
-      end
-    end
-  end
-
-  context "user visits a non-admin page after the timeout period" do
-    let(:after_expiry) { timeout_length_in_minutes.minutes + 1.second }
-
-    it "still clears the admin session" do
-      expect(session[:user_id]).to eq(user_id)
-      expect(session[:organisation_id]).to eq(organisation_id)
-      expect(session[:role_codes]).to eq([AdminSession::SERVICE_OPERATOR_DFE_SIGN_IN_ROLE_CODE])
-
-      travel after_expiry do
-        get new_claim_path
-
-        expect(session[:user_id]).to be_nil
-        expect(session[:organisation_id]).to be_nil
-        expect(session[:role_codes]).to be_nil
-
-        expect(response).to be_successful
       end
     end
   end